refactor(frontend): do all load actions on page level tm-94
