--- apps/frontend/src/pages/friends/components/components.ts
@@ -1,3 +1,0 @@
-export { Followers } from "./followers/followers.js";
-export { Followings } from "./followings/followings.js";
-export { PotentialFriends } from "./potential-friends/potential-friends.js";
