+++ apps/frontend/src/pages/friends/libs/components/components.ts
@@ -1,0 +1,1 @@
+export { FriendList } from "./friend-list/friend-list.js";
