+++ apps/frontend/src/pages/friends/libs/components/friend-list/friend-list.tsx
