+++ apps/frontend/src/pages/friends/libs/components/friend-list/styles.module.css
