+++ apps/frontend/src/pages/friends/libs/components/friend/friend.tsx
