+++ apps/frontend/src/pages/friends/libs/components/friend/styles.module.css
